separate out pool from buffer file
