+++ pkg/writers/buffered_file_writer/bufferedfilewriter.go
@@ -18,7 +18,9 @@ import (
 // This allows for efficient reuse of buffers across multiple writers.
 var sharedBufferPool *buffer.Pool
 
-func init() { sharedBufferPool = buffer.NewBufferPool() }
+func init() {
+	sharedBufferPool = buffer.NewBufferPool()
+}
 
 type bufferedFileWriterMetrics struct{}
 
@@ -54,7 +56,7 @@ type BufferedFileWriter struct {
 	size      uint64 // Total size of the data written.
 
 	bufPool  *buffer.Pool   // Pool for storing buffers for reuse.
-	buf      *buffer.Buffer // Buffer for storing data under the threshold in memory.
+	buf      *buffer.Ring   // Buffer for storing data under the threshold in memory.
 	filename string         // Name of the temporary file.
 	file     io.WriteCloser // File for storing data over the threshold.
 
@@ -76,7 +78,7 @@ const defaultThreshold = 10 * 1024 * 1024 // 10MB
 func New(ctx context.Context, opts ...Option) *BufferedFileWriter {
 	buf := sharedBufferPool.Get(ctx)
 	if buf == nil {
-		buf = buffer.NewBuffer()
+		buf = buffer.NewRingBuffer(1 << 12)
 	}
 	w := &BufferedFileWriter{
 		threshold: defaultThreshold,
@@ -97,7 +99,7 @@ func (w *BufferedFileWriter) Len() int { return int(w.size) }
 // String returns all the data written to the buffer or file as a string or an empty string if there is an error.
 func (w *BufferedFileWriter) String() (string, error) {
 	if w.file == nil {
-		return w.buf.String(), nil
+		return string(w.buf.Bytes()), nil
 	}
 
 	// Data is in a file, read from the file.
